Moved old macOS software update table impl to proper package.
