+++ pkg/osquery/tables/macos_software_update/SUSharedPrefs.h
