+++ pkg/osquery/tables/macos_software_update/sus_darwin.go
@@ -1,4 +1,4 @@
-package table
+package macos_software_update
 
 /*
 #cgo darwin CFLAGS: -DDARWIN -x objective-c
